catalyst: Make and use squashfs snapshots
There were a number of problems with catalyst's snapshot system. It was
built around using the build system's portdir and had no control over
what was in that portdir or when it was updated.

As a result, when a stage build failed, it was difficult to tell whether
the snapshot contained a particular recent fix or not.

Without snapcache (which had its own set of problems), ebuild repo
snapshots were tar'd and compressed and then unpacked into the stage
chroot which is an unnecessarily expensive process. Moreover, a porttree
has more than 100k small files, which are stored extremely inefficiently
on most file systems—a whole porttree is usually around 700M on disk.

Instead, we make a compressed squashfs image and mount it in the build
chroot. squashfs deduplicates files, which makes a large difference in
our usage and then compresses. A squashfs image of the porttree with
gzip -9 compression is ~38M.

Using git archive and tar2sqfs the squashfs image can be generated from
a bare git repo in less than 10 seconds on a modern system. The bare git
repo can be fetched with --depth=1 to minimize bandwidth usage and
storage requirements (which is what this code does). Storage
requirements for the git porttree with metadata are ~70M.

When the snapshot target is run, it automatically updates and garbage
collects the porttree git repo before making and storing a squashfs
image in /var/tmp/catalyst/snapshots/. The filename is

Signed-off-by: Matt Turner <mattst88@gentoo.org>